backend/requirements.txt
fastapi
uvicorn[standard]
motor
pyjwt
bcrypt
python-dotenv
pydantic-settings
email-validator
python-multipart
pymupdf
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-pinecone
langchain-google-genai
sentence-transformers
pinecone-client
google-generativeai
reportlab
python-docx
redis
